crud with authentication

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel.js');
+const User = require('../models/userModel.js');
 
 //Get goal
 //@route GET /api/goals
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -19,6 +20,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -27,14 +29,27 @@ const setGoal = asyncHandler(async (req, res) => {
 //Update goal
 //@route PUT /api/goals/:id
 const updateGoal = asyncHandler(async (req, res) => {
-
   const goal = await Goal.findById(req.params.id);
 
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
   }
-  
+
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  //the  logged user must match the goal user
+  if (goal.user.toString() != user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -52,9 +67,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error('Goal not found');
   }
 
-  console.log('goal', goal);
+  const user = await User.findById(req.user.id);
 
-  await goal.deleteOne();//! .remove() depricated
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  //the  logged user must match the goal user
+  if (goal.user.toString() != user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
+  await goal.deleteOne(); //! .remove() depricated
 
   res.status(200).json({ id: req.params.id });
 });
